Add files via upload
I found a guide on how to make a cashier system in this video, https://www.youtube.com/watch?v=0pqQjXpifXo. It is currently using a menu for a pizza place, so we will have to customize it to make it an ice cream menu. It also has NPCs to order the food so it is a good starting point.