Add files via upload
These are the models that I want to use in the Ice Cream Store to distribute the ice cream. The case has a lot of different options so it will give players more options and the soft serve machine is another option. We can put these in front of and behind the counter.