Ice Cream Case in the Store
I added ice cream cases in front of the cash registers to represent where the cashiers get the ice cream from. In the future, we can add the ability to get ice cream from the cases. Next week I want to continue to change the store into an ice cream parlor.