New Color Scheme on the Building
I changed all of the red to pink so it looks more like an ice cream parlor. I couldn't figure out how to not make the CPUs climb onto the ice cream case, however. I need to do more research into how to fix that next week. I also went through and sorted all of the models in the world so that it is easier to change the right things.